// types.ts
export interface NavItem {
  label: string;
  path: string;
}

export interface RoomPackage {
  title: string;
  price: string;
  image: string;
  description: string;
}

export interface Activity {
  title: string;
  image: string;
}

export interface Destination {
  title: string;
  description: string;
  image: string;
}